be/requirements.txt
uvicorn
fastapi
scrapy
requests
schedule
beautifulsoup4
aiohttp
python-crontab
sqlalchemy
mysql-connector-python
passlib[bcrypt]
pydantic
mysqlclient
PyJWT
croniter
googletrans==4.0.0-rc1